requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.5.0
django-leaflet==0.27.1
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-stdimage==5.1.1
django-userforeignkey==0.3.0
djangorestframework==3.12.1
djangorestframework-gis==0.16
djangorestframework-simplejwt==4.6.0
drf-base64==2.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
progressbar2==3.53.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
